+++ src/main/kotlin/io/github/andrepg/flatpak/runs/execution/commands/CustomCommandFactory.kt
@@ -5,8 +5,8 @@ import io.github.andrepg.flatpak.runs.configuration.FlatpakRunSettings
 class CustomCommandFactory : CommandFactory() {
     override fun create(settings: FlatpakRunSettings): List<String> = this.getFlatpakCommand().plus(
         listOf(
-            settings.buildDir,
-            settings.manifestPath,
+            settings.effectiveBuildDir(),
+            settings.effectiveManifestPath(),
             settings.customArguments.joinToString(" ")
         )
     )