+++ src/main/kotlin/io/github/andrepg/flatpak/runs/execution/commands/ValidateManifestCommandFactory.kt
@@ -4,5 +4,5 @@ import io.github.andrepg.flatpak.runs.configuration.FlatpakRunSettings
 
 class ValidateManifestCommandFactory : CommandFactory() {
     override fun create(settings: FlatpakRunSettings): List<String> =
-        this.getFlatpakCommand().plus(listOf("--show-manifest", settings.manifestPath))
+        this.getFlatpakCommand().plus(listOf("--show-manifest", settings.effectiveManifestPath()))
 }